Set event timezone to Pacific Standard Time (PST)
Apple always announces their event times in the PST time zone. However, in `main.js`, the time zone is hardcoded to `GMT-0700`, which can lead to wrong event times because daylight saving time is not automatically respected (for example in Germany, the website currently says "18:00" for the "One more thing" event, which should be "19:00"). I changed the time zone to PST, which should fix the issue.